Fixed #377 - Moved the object builder adapter to the core-impl and removed the QueryTransformer
