+++ core/impl/src/main/java/com/blazebit/persistence/impl/AbstractCTECriteriaBuilder.java
@@ -23,6 +23,7 @@ import com.blazebit.persistence.impl.expression.NullExpression;
 import com.blazebit.persistence.impl.expression.PathExpression;
 import com.blazebit.persistence.impl.expression.PropertyExpression;
 import com.blazebit.persistence.impl.query.CTEQuerySpecification;
+import com.blazebit.persistence.impl.query.CustomSQLQuery;
 import com.blazebit.persistence.impl.query.EntityFunctionNode;
 import com.blazebit.persistence.impl.query.QuerySpecification;
 import com.blazebit.persistence.spi.DbmsStatementType;
