+++ core/impl/src/main/java/com/blazebit/persistence/impl/CriteriaBuilderConfigurationImpl.java
@@ -139,7 +139,6 @@ import com.blazebit.persistence.spi.EntityManagerFactoryIntegrator;
 import com.blazebit.persistence.spi.ExtendedQuerySupport;
 import com.blazebit.persistence.spi.JpqlFunctionGroup;
 import com.blazebit.persistence.spi.JpqlMacro;
-import com.blazebit.persistence.spi.QueryTransformer;
 import com.blazebit.persistence.spi.SetOperationType;
 
 /**
@@ -149,7 +148,6 @@ import com.blazebit.persistence.spi.SetOperationType;
  */
 public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfiguration {
 
-    private final List<QueryTransformer> queryTransformers = new ArrayList<QueryTransformer>();
     private final Map<String, DbmsDialect> dbmsDialects = new HashMap<String, DbmsDialect>();
     private final Map<String, JpqlFunctionGroup> functions = new HashMap<String, JpqlFunctionGroup>();
     private final Map<String, Class<?>> treatTypes = new HashMap<String, Class<?>>();
@@ -160,7 +158,6 @@ public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfigur
 
     public CriteriaBuilderConfigurationImpl() {
         loadDefaultProperties();
-        loadQueryTransformers();
         loadExtendedQuerySupport();
         loadEntityManagerIntegrator();
         loadDbmsDialects();
@@ -438,16 +435,6 @@ public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfigur
         properties.put(ConfigurationProperties.EXPRESSION_CACHE_CLASS, ConcurrentHashMapExpressionCache.class.getName());
     }
 
-    private void loadQueryTransformers() {
-        ServiceLoader<QueryTransformer> serviceLoader = ServiceLoader.load(QueryTransformer.class);
-        Iterator<QueryTransformer> iterator = serviceLoader.iterator();
-
-        if (iterator.hasNext()) {
-            QueryTransformer transformer = iterator.next();
-            queryTransformers.add(transformer);
-        }
-    }
-
     private void loadExtendedQuerySupport() {
         ServiceLoader<ExtendedQuerySupport> serviceLoader = ServiceLoader.load(ExtendedQuerySupport.class);
         Iterator<ExtendedQuerySupport> iterator = serviceLoader.iterator();
@@ -517,17 +504,6 @@ public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfigur
         return dbmsDialects;
     }
 
-    @Override
-    public CriteriaBuilderConfiguration registerQueryTransformer(QueryTransformer transformer) {
-        queryTransformers.add(transformer);
-        return this;
-    }
-
-    @Override
-    public List<QueryTransformer> getQueryTransformers() {
-        return queryTransformers;
-    }
-    
     public ExtendedQuerySupport getExtendedQuerySupport() {
         return extendedQuerySupport;
     }
